+++ supplier/sms/app/controllers/YHDGroupBuy.java
@@ -236,7 +236,7 @@ public class YHDGroupBuy extends Controller {
 
     // 创建一百券订单
     private static Order createYbqOrder(String outerGroupId, BigDecimal productPrize,
-                                        Integer productNum, String userPhone, List<YHDErrorInfo> errorInfoList) {
+                                        Long productNum, String userPhone, List<YHDErrorInfo> errorInfoList) {
         Resaler resaler = Resaler.findOneByLoginName(YHD_LOGIN_NAME);
         if (resaler == null) {
             Logger.error("can not find the resaler by login name: %s", YHD_LOGIN_NAME);
